Be

See merge request s09-webmobile1-sub2/S09P12E201!34

Former-commit-id: a9b7990e8e135e69e1d80d82766657958a300c04

+++ backend/src/main/java/com/ssafy/api/service/MyTrophyRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.api.model.MyTrophy;
 
 @Repository
-public interface MyTrophyRepository extends JpaRepository<MyTrophy, Integer> {
+public interface MyTrophyRepository extends JpaRepository<MyTrophy, Long> {
 	
-	void deleteByUserID(int userID);
+	void deleteByUserID(long userID);
 
-	List<MyTrophy> findAllByUserID(int userID);
+	List<MyTrophy> findAllByUserID(long userID);
 
 }
